// artifacts/grace-nepal/src/lib/types.ts
export interface LoginRequest {
  email: string
  password: string
}